+++ app/src/main/java/com/sekusarisu/mdpings/vpings/data/networking/dto/DashboardServerDto.kt
@@ -20,40 +20,40 @@ data class DashboardServerDto(
 @Serializable
 data class Host(
     val platform: String = "unknown",
-    @SerialName(value = "platform_version") val platformVersion: String,
-    val cpu: List<String>,
-    @SerialName(value = "mem_total") val memTotal: Long,
-    @SerialName(value = "disk_total") val diskTotal: Long,
+    @SerialName(value = "platform_version") val platformVersion: String = "unknown",
+    val cpu: List<String> = listOf<String>("unknown"),
+    @SerialName(value = "mem_total") val memTotal: Long = 0,
+    @SerialName(value = "disk_total") val diskTotal: Long = 0,
     @SerialName(value = "swap_total") val swapTotal: Long = 0,
     val arch: String = "unknown",
     val virtualization: String = "unknown",
-    @SerialName(value = "boot_time") val bootTime: Long,
+    @SerialName(value = "boot_time") val bootTime: Long = 0,
     val version: String = "unknown"
 )
 
 @Serializable
 data class State(
     val cpu: Double = 0.0,
-    @SerialName(value = "mem_used") val memUsed: Long,
-    @SerialName(value = "disk_used") val diskUsed: Long,
-    @SerialName(value = "net_in_transfer") val netInTransfer: Long,
-    @SerialName(value = "net_out_transfer") val netOutTransfer: Long,
-    @SerialName(value = "net_in_speed") val netInSpeed: Long,
-    @SerialName(value = "net_out_speed") val netOutSpeed: Long,
-    val uptime: Long,
-    @SerialName(value = "tcp_conn_count") val tcpConnCount: Long,
-    @SerialName(value = "udp_conn_count") val udpConnCount: Long,
-    @SerialName(value = "process_count") val processCount: Long
+    @SerialName(value = "mem_used") val memUsed: Long = 0,
+    @SerialName(value = "disk_used") val diskUsed: Long = 0,
+    @SerialName(value = "net_in_transfer") val netInTransfer: Long = 0,
+    @SerialName(value = "net_out_transfer") val netOutTransfer: Long = 0,
+    @SerialName(value = "net_in_speed") val netInSpeed: Long = 0,
+    @SerialName(value = "net_out_speed") val netOutSpeed: Long = 0,
+    val uptime: Long = 0,
+    @SerialName(value = "tcp_conn_count") val tcpConnCount: Long = 0,
+    @SerialName(value = "udp_conn_count") val udpConnCount: Long = 0,
+    @SerialName(value = "process_count") val processCount: Long = 0
 )
 
 @Serializable
 data class Geoip(
     val ip: Ip,
-    @SerialName(value = "country_code") val countryCode: String
+    @SerialName(value = "country_code") val countryCode: String = "un"
 )
 
 @Serializable
 data class Ip(
-    @SerialName(value = "ipv4_addr") val ipv4Addr: String,
+    @SerialName(value = "ipv4_addr") val ipv4Addr: String = "1.1.1.1",
     @SerialName(value = "ipv6_addr") val ipv6Addr: String = "N/A"
 )