+++ app/src/main/java/com/sekusarisu/mdpings/vpings/domain/SessionData.kt
@@ -1,0 +1,7 @@
+package com.sekusarisu.mdpings.vpings.domain
+
+data class SessionData(
+    val sessionId: String,
+    val serverId: Int,
+    val serverName: String
+)