+++ app/src/main/java/com/sekusarisu/mdpings/vpings/presentation/server_terminal/ServerTerminalState.kt
@@ -1,0 +1,11 @@
+package com.sekusarisu.mdpings.vpings.presentation.server_terminal
+
+import androidx.compose.runtime.Immutable
+
+
+@Immutable
+data class ServerTerminalState(
+    val isLoading: Boolean = false,
+    val selectedServerId: Int = 0,
+    val terminal: String = ""
+)